Added upgrade status put request and balance calculation
Once a user can do an status upgrade. The request will update the user's status id to the next one and also change his corresponding status information. For to total balance. each time the get user by email is pulled for the overview it will update the balance to ensure it is always up to date.